Inline Case files for IteratorHasNextCallsNextMethodSpec
